requirements.txt
# Core Dependencies
faiss-cpu>=1.7.4
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
numpy>=1.24.0

# Backend
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data & Storage
sqlite-utils>=3.34.0
pandas>=2.0.0

# Code Analysis
tree-sitter>=0.20.0
tree-sitter-languages>=1.0.0

# CLI & Utils
click>=8.0.0
rich>=13.0.0
tqdm>=4.65.0

# Optimization
onnxruntime>=1.15.0
optimum>=1.10.0

# Monitoring & Observability
prometheus-client>=0.17.0
psutil>=5.9.0

# Security
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
mypy>=1.0.0
flake8>=6.0.0
types-requests>=2.31.0
types-setuptools>=68.0.0
httpx>=0.25.0
